refactor: move config module up
