+++ packages/apps/reputation-oracle/server/src/app.module.ts
@@ -9,13 +9,13 @@ import { HttpValidationPipe } from './common/pipes';
 import { HealthModule } from './modules/health/health.module';
 import { ReputationModule } from './modules/reputation/reputation.module';
 import { Web3Module } from './modules/web3/web3.module';
-import { envValidator } from './common/config';
+import { envValidator } from './config';
 import { AuthModule } from './modules/auth/auth.module';
 import { TransformInterceptor } from './common/interceptors/transform.interceptor';
 import { KycModule } from './modules/kyc/kyc.module';
 import { CronJobModule } from './modules/cron-job/cron-job.module';
 import { PayoutModule } from './modules/payout/payout.module';
-import { EnvConfigModule } from './common/config/config.module';
+import { EnvConfigModule } from './config/config.module';
 import { HCaptchaModule } from './integrations/hcaptcha/hcaptcha.module';
 import { ExceptionFilter } from './common/filters/exception.filter';
 import { QualificationModule } from './modules/qualification/qualification.module';
