+++ packages/apps/reputation-oracle/server/src/modules/escrow-completion/escrow-completion.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { BaseRepository } from '../../database/base.repository';
 import { DataSource, LessThanOrEqual } from 'typeorm';
-import { ServerConfigService } from '../../common/config/server-config.service';
+import { ServerConfigService } from '../../config/server-config.service';
 import { EscrowCompletionEntity } from './escrow-completion.entity';
 import { EscrowCompletionStatus } from '../../common/enums';
