+++ packages/apps/reputation-oracle/server/src/modules/kyc/kyc.service.spec.ts
@@ -6,9 +6,9 @@ import { Test } from '@nestjs/testing';
 import { of } from 'rxjs';
 import { DeepPartial } from 'typeorm';
 import { MOCK_ADDRESS, mockConfig } from '../../../test/constants';
-import { HCaptchaConfigService } from '../../common/config/hcaptcha-config.service';
-import { NetworkConfigService } from '../../common/config/network-config.service';
-import { KycConfigService } from '../../common/config/kyc-config.service';
+import { HCaptchaConfigService } from '../../config/hcaptcha-config.service';
+import { NetworkConfigService } from '../../config/network-config.service';
+import { KycConfigService } from '../../config/kyc-config.service';
 import { KycStatus } from '../../common/enums/user';
 import { Web3Service } from '../web3/web3.service';
 import { KycEntity } from './kyc.entity';
