+++ packages/apps/reputation-oracle/server/src/modules/storage/storage.service.ts
@@ -12,8 +12,8 @@ import crypto from 'crypto';
 import { UploadedFile } from '../../common/interfaces/s3';
 import { Web3Service } from '../web3/web3.service';
 import { FortuneFinalResult } from '../../common/interfaces/job-result';
-import { S3ConfigService } from '../../common/config/s3-config.service';
-import { PGPConfigService } from '../../common/config/pgp-config.service';
+import { S3ConfigService } from '../../config/s3-config.service';
+import { PGPConfigService } from '../../config/pgp-config.service';
 import { isNotFoundError } from '../../common/errors/minio';
 import {
   FileDownloadError,
