+++ packages/apps/reputation-oracle/server/src/modules/user/user.service.ts
@@ -13,13 +13,13 @@ import { UserRepository } from './user.repository';
 import { Web3Service } from '../web3/web3.service';
 import { SignatureType, Web3Env } from '../../common/enums/web3';
 import { ChainId, KVStoreClient, KVStoreUtils } from '@human-protocol/sdk';
-import { Web3ConfigService } from '../../common/config/web3-config.service';
+import { Web3ConfigService } from '../../config/web3-config.service';
 import { SiteKeyEntity } from './site-key.entity';
 import { SiteKeyRepository } from './site-key.repository';
 import { SiteKeyType } from '../../common/enums';
 import { HCaptchaService } from '../../integrations/hcaptcha/hcaptcha.service';
-import { HCaptchaConfigService } from '../../common/config/hcaptcha-config.service';
-import { NetworkConfigService } from '../../common/config/network-config.service';
+import { HCaptchaConfigService } from '../../config/hcaptcha-config.service';
+import { NetworkConfigService } from '../../config/network-config.service';
 import { prepareSignatureBody } from '../../utils/web3';
 import { KycSignedAddressDto } from '../kyc/kyc.dto';
 import { ethers } from 'ethers';
