+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook-incoming.service.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { Injectable } from '@nestjs/common';
-import { ServerConfigService } from '../../common/config/server-config.service';
-import { Web3ConfigService } from '../../common/config/web3-config.service';
+import { ServerConfigService } from '../../config/server-config.service';
+import { Web3ConfigService } from '../../config/web3-config.service';
 import { BACKOFF_INTERVAL_SECONDS } from '../../common/constants';
 import { EventType, WebhookIncomingStatus } from '../../common/enums';
 import { calculateExponentialBackoffMs } from '../../utils/backoff';
