+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook-outgoing.repository.ts
@@ -3,7 +3,7 @@ import { BaseRepository } from '../../database/base.repository';
 import { DataSource, LessThanOrEqual } from 'typeorm';
 import { WebhookOutgoingStatus } from '../../common/enums/webhook';
 import { WebhookOutgoingEntity } from './webhook-outgoing.entity';
-import { ServerConfigService } from '../../common/config/server-config.service';
+import { ServerConfigService } from '../../config/server-config.service';
 
 @Injectable()
 export class WebhookOutgoingRepository extends BaseRepository<WebhookOutgoingEntity> {
