+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook-outgoing.service.spec.ts
@@ -23,8 +23,8 @@ import { of } from 'rxjs';
 import { HEADER_SIGNATURE_KEY } from '../../common/constants';
 import { signMessage } from '../../utils/web3';
 import { HttpStatus } from '@nestjs/common';
-import { Web3ConfigService } from '../../common/config/web3-config.service';
-import { ServerConfigService } from '../../common/config/server-config.service';
+import { Web3ConfigService } from '../../config/web3-config.service';
+import { ServerConfigService } from '../../config/server-config.service';
 import { OutgoingWebhookError, WebhookErrorMessage } from './webhook.error';
 
 describe('WebhookOutgoingService', () => {
